Change of layour for functions. Also new example notebook.
