+++ docs/source/conf.py
@@ -9,7 +9,7 @@
 project = "mff"
 copyright = "2024, IMF"
 author = "Ando Sakai, Doga Bilgin and Sultan Orazbayev"
-release = "0.1"
+release = "0.1.2"
 
 
 # -- General configuration ---------------------------------------------------
