+++ docs/source/mff_documentation.rst
@@ -1,8 +1,19 @@
 Macroframework Forecasting Package
 ----------------------------------
 
-.. automodule:: mff.mff
+.. automodule:: mff.MFF
    :members:
    :member-order: groupwise
    :show-inheritance:
+
+.. automodule:: mff.MFF_mixed_frequency
+   :members:
+   :member-order: groupwise
+   :show-inheritance:
+
+.. automodule:: mff.utils
+   :members:
+   :member-order: groupwise
+   :show-inheritance:
+
